Switch to int gauge

+++ src/http/websockets.rs
@@ -455,7 +455,7 @@ async fn handle_ws_client(
 
     let cid = client.get_client_id().clone();
     state.clients.lock().await.insert(client_id.clone(), client);
-    let total = state.clients.lock().await.len() as f64;
+    let total = state.clients.lock().await.len() as i64;
     TOTAL_CLIENTS.get().unwrap().set(total);
 
     debug!("client {client_id} was connected");
@@ -513,7 +513,7 @@ async fn handle_ws_client(
         }
     }
 
-    let total_clients = state.clients.lock().await.len() as f64;
+    let total_clients = state.clients.lock().await.len() as i64;
     TOTAL_CLIENTS.get().unwrap().set(total_clients);
     debug!("total connected clients: {}", total_clients);
 }
